Apply final project improvements

+++ src/main/java/com/cfg/petadoption/entity/Pet.java
@@ -3,6 +3,7 @@ package com.cfg.petadoption.entity;
 import jakarta.persistence.*;
 import lombok.Data;
 import lombok.NoArgsConstructor;
+
 // Represents a pet stored in the database.
 @Entity
 @Table(name = "pets")
@@ -13,7 +14,7 @@ public class Pet {
     // Automatically generated primary key.
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private long id;
+    private Integer id;
     private String name;
     private String species;
     private String breed;
